Rename Heritage to Story
