+++ src/components/Story/Story.vue
@@ -1,0 +1,11 @@
+<template>
+
+</template>
+
+<script>
+export default {
+};
+</script>
+
+<style scoped>
+</style>
