+++ src/pages/Album.vue
