+++ src/pages/Story.vue
